* 'current' of github.com:vyos/vyos-1x:
  T1749: support multiple ranges in the numeric validator.
  dhcp-server: T1745: bugfix corner case on static-assignments
  system-proxy: T1741 - Add system wide proxy setting
  wireguard - remove endpoint check to enable roaming connections
  system-proxy: T1741 - Add system wide proxy setting CLI implementation
  Python/ifconfig: T1712: always start DHCP when configured
  Python/ifconfig: T1557: get_status() must use admin state not operstate
  bgp: T1490: fix migrator file permissions
  snmp: T1737: add missing completion helpers
  Revert "Python/ifconfig: T1712: wait when changing interface state"
  snmpd: T1705 - High CPU usage by bgpd when snmp is active
  Revert "snmpd: T1705 - High CPU usage by bgpd when snmp is active"
  openvpn: T1548: clean out import statements
  ssh.py: check if file exists before deleting it
  [BGP] T1490: Added migration for obsoleted 'bgp scan-time' parameter

+++ src/conf_mode/ssh.py
@@ -262,7 +262,8 @@ def apply(ssh):
     else:
         # SSH access is removed in the commit
         os.system("sudo systemctl stop ssh.service")
-        os.unlink(config_file)
+        if os.path.isfile(config_file):
+            os.unlink(config_file)
 
     return None
